be able to use up durability on tool

+++ roborenewal/Assets/Scripts/Tools/Tool.cs
@@ -10,7 +10,7 @@ public abstract class Tool : MonoBehaviour
     public int mDurability = 10;
 
     public float mAttackRate = 1.5f;
-    public int mDamage = 1;
+    public float mDamage = 1;
 
     public Action OnToolBroken;
